% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-priority-queue.R
\name{cpq_create}
\alias{cpq_create}
\alias{cpq_length}
\alias{cpq_is_empty}
\alias{cpq_insert}
\alias{cpq_pop}
\alias{cpq_peek}
\alias{cpq_dump}
\title{My Priority Queue}
\usage{
cpq_create()

cpq_length(cpq)

cpq_is_empty(cpq)

cpq_insert(cpq, obj, priority)

cpq_pop(cpq)

cpq_peek(cpq)

cpq_length(cpq)

cpq_dump(cpq)
}
\arguments{
\item{cpq}{my priority queue object}

\item{obj}{any R object}

\item{priority}{numeric priority (must be finite and not NA or NaN)}
}
\description{
My Priority Queue
}
