% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-priority-queue.R
\name{rpq_create}
\alias{rpq_create}
\alias{rpq_insert}
\alias{rpq_pop}
\alias{rpq_peek}
\alias{rpq_is_empty}
\alias{rpq_length}
\title{A max priority queue for R objects (implemented in R)}
\usage{
rpq_create()

rpq_insert(rpq, obj, priority)

rpq_pop(rpq)

rpq_peek(rpq)

rpq_is_empty(rpq)

rpq_length(rpq)
}
\arguments{
\item{rpq}{priority queue object as created by \code{rpq_create()}}

\item{obj}{Any R object}

\item{priority}{numeric prioirty}
}
\description{
A max priority queue for R objects (implemented in R)
}
